clean up and refactoring
